+++ NNLS_XXZ.jl
@@ -21,7 +21,7 @@ settings = ArgParseSettings(prog="NNLS code for XXZ model"
         default = 10.0
         help = "β: inverse temperature"
     "--D"
-        arg_type = Integer
+        arg_type = Int64
         default = 8
         help = "cMPO bond dimension"
     "--operator"
